Vacuum model and prelim line renderer adjustment
Added a vacuum looking model to the left hand, fixed the line renderer to increase the spread of the radius that it reaches.
Line renderer solely used to show area of effect for the vacuum, for actual sucking, create an animation and play it on loop while trigger held down.