Radio script adjusted to work with less text fields
Radio script now should work when theres less text fields that should be displayed. This way you could reuse this script on a different game object if you had to display varying information, useful for shop items, or something else.
If reusing this script, should take into consideration that onPickUp should not be used, as it moves the text fields to a static position. If wanting to use onPickUp, should introduce new variables that would handle where the text fields go when the object is picked up. ☺